Move optimisation checkpointing case to separate demo
This should keep the adjoint directory a bit cleaner.
